+++ src/tools/ToolsClassic/vmap_tools/vmap4_extractor/gameobject_extract.cpp
@@ -111,7 +111,7 @@ void ExtractGameobjectModels()
         if (result)
         {
             uint32 displayId = it->getUInt(0);
-            uint32 path_length = strlen(name);
+            uint32 path_length = static_cast<uint32_t>(strlen(name));
             fwrite(&displayId, sizeof(uint32), 1, model_list);
             fwrite(&path_length, sizeof(uint32), 1, model_list);
             fwrite(name, sizeof(char), path_length, model_list);
