+++ src/tools/ToolsMop/vmap_tools/vmap4_extractor/adtfile.cpp
@@ -210,7 +210,7 @@ bool ADTFile::init(uint32 map_num, uint32 tileX, uint32 tileY)
         }
 
         //======================
-        ADT.seek(nextpos);
+        ADT.seek(static_cast<int>(nextpos));
     }
 
     ADT.close();
