+++ src/tools/ToolsTBC/mmaps_generator/TerrainBuilder.cpp
@@ -187,11 +187,11 @@ namespace MMAP
             {
                 uint8 v9[V9_SIZE_SQ];
                 uint8 v8[V8_SIZE_SQ];
-                size_t count = 0;
-                count += fread(v9, sizeof(uint8), V9_SIZE_SQ, mapFile);
-                count += fread(v8, sizeof(uint8), V8_SIZE_SQ, mapFile);
+                int count = 0;
+                count += static_cast<int>(fread(v9, sizeof(uint8), V9_SIZE_SQ, mapFile));
+                count += static_cast<int>(fread(v8, sizeof(uint8), V8_SIZE_SQ, mapFile));
                 if (count != expected)
-                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %llu\n", expected, count);
+                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %d\n", expected, count);
 
                 heightMultiplier = (hheader.gridMaxHeight - hheader.gridHeight) / 255;
 
@@ -205,11 +205,11 @@ namespace MMAP
             {
                 uint16 v9[V9_SIZE_SQ];
                 uint16 v8[V8_SIZE_SQ];
-                size_t count = 0;
-                count += fread(v9, sizeof(uint16), V9_SIZE_SQ, mapFile);
-                count += fread(v8, sizeof(uint16), V8_SIZE_SQ, mapFile);
+                int count = 0;
+                count += static_cast<int>(fread(v9, sizeof(uint16), V9_SIZE_SQ, mapFile));
+                count += static_cast<int>(fread(v8, sizeof(uint16), V8_SIZE_SQ, mapFile));
                 if (count != expected)
-                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %llu\n", expected, count);
+                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %d\n", expected, count);
 
                 heightMultiplier = (hheader.gridMaxHeight - hheader.gridHeight) / 65535;
 
@@ -221,11 +221,11 @@ namespace MMAP
             }
             else
             {
-                size_t count = 0;
-                count += fread(V9, sizeof(float), V9_SIZE_SQ, mapFile);
-                count += fread(V8, sizeof(float), V8_SIZE_SQ, mapFile);
+                int count = 0;
+                count += static_cast<int>(fread(V9, sizeof(float), V9_SIZE_SQ, mapFile));
+                count += static_cast<int>(fread(V8, sizeof(float), V8_SIZE_SQ, mapFile));
                 if (count != expected)
-                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %llu\n", expected, count);
+                    printf("TerrainBuilder::loadMap: Failed to read some data expected %d, read %d\n", expected, count);
             }
 
             // hole data
